+++ includes/draw.h
@@ -6,13 +6,14 @@
 /*   By: tsishika <tsishika@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/12 16:11:24 by tsishika          #+#    #+#             */
-/*   Updated: 2024/03/20 01:22:41 by tsishika         ###   ########.fr       */
+/*   Updated: 2024/03/20 01:34:02 by tsishika         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef DRAW_H
 # define DRAW_H
 
+# include "ray.h"
 # include "hook.h"
 # include "struct.h"
 # include "utils.h"
@@ -25,6 +26,7 @@
 # define PLAYER_COLOR 0xFF0000
 # define BACKGROUND 0xFFFFFF
 # define SCALE 10
+# define PIXEL 64
 
 // draw_texture.c
 unsigned int	*get_pixel_color(t_texture *texture, int y, int x);
@@ -50,7 +52,7 @@ void			init_ray(t_data *data, t_ray *ray, int x);
 void			calculate_step_and_side_dist(t_data *data, t_ray *ray);
 void			perform_dda(t_data *data, t_ray *ray);
 void			calculate_wall_height(t_data *data, t_ray *ray, t_wall *wall);
-void			draw_wall(t_data *data, t_ray *ray, t_wall *wall, int x);
+void			draw_wall(t_data *data, t_wall *wall, int x);
 
 // xpm_file_to_img.c
 void			xpm_file_to_img(t_data *data,
